backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
httpx
python-dotenv
pydantic
